# homework/server/docs/file_transfer.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "file_transfer.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "FileTransferService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/files": {
      "get": {
        "operationId": "FileTransferService_GetFileList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/file_transfer_grpcGetFileListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "FileTransferService"
        ]
      }
    },
    "/files/{name}/content": {
      "get": {
        "operationId": "FileTransferService_GetFile",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/file_transfer_grpcFileChunk"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of file_transfer_grpcFileChunk"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "FileTransferService"
        ]
      }
    },
    "/files/{name}/info": {
      "get": {
        "operationId": "FileTransferService_GetFileInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/file_transfer_grpcGetFileInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "FileTransferService"
        ]
      }
    }
  },
  "definitions": {
    "file_transfer_grpcFileChunk": {
      "type": "object",
      "properties": {
        "chunkData": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "file_transfer_grpcGetFileInfoResponse": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "size": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "creationDate": {
          "type": "string"
        },
        "lastModifiedDate": {
          "type": "string"
        },
        "accessRights": {
          "type": "string"
        },
        "location": {
          "type": "string"
        }
      }
    },
    "file_transfer_grpcGetFileListResponse": {
      "type": "object",
      "properties": {
        "names": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}